extracts ETagAndResponseWrapper
